requirements.txt
fastapi==0.115.2
uvicorn==0.29.0
httpx==0.27.0
requests==2.31.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
prophet==1.1.5
pystan==2.19.1.1
python-dateutil==2.8.2
tqdm==4.66.2
python-dotenv==1.0.1
